% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apdf.R
\name{add_apdf_date}
\alias{add_apdf_date}
\title{Add the Operating Date to APDF Data}
\usage{
add_apdf_date(apdf)
}
\arguments{
\item{apdf}{A decoded APDF tibble.}
}
\value{
A tibble.
}
\description{
Adds a \code{DATE} column using \code{BLOCK_TIME} when available and \code{MVT_TIME}
otherwise.
}
