% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode.R
\name{apdf_dictionary}
\alias{apdf_dictionary}
\title{APDF Dictionary}
\usage{
apdf_dictionary()
}
\value{
A tibble with \code{SOURCE_NAME}, \code{TARGET_NAME}, and \code{DESCRIPTION}.
}
\description{
Returns a template dictionary used to map source column names from APDF
extracts to harmonised package names.
}
