% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apdf_tma.R
\name{apdf_tma_fields}
\alias{apdf_tma_fields}
\title{APDF Terminal-Airspace Reference Fields}
\usage{
apdf_tma_fields()
}
\value{
A character vector of source field names.
}
\description{
Returns the default fields used for terminal-airspace reference preparation.
}
