% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apdf.R
\name{apdf_traffic_fields}
\alias{apdf_traffic_fields}
\title{APDF Traffic Dictionary}
\usage{
apdf_traffic_fields()
}
\value{
A character vector of source field names.
}
\description{
Returns the default fields used for airport traffic preparation.
}
