% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apdf_txxt.R
\name{apdf_txxt_fields}
\alias{apdf_txxt_fields}
\title{APDF Taxi-Time Reference Fields}
\usage{
apdf_txxt_fields()
}
\value{
A character vector of source field names.
}
\description{
Returns the default fields used for taxi-time reference preparation.
}
