% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apdf_tma.R
\name{build_tma_sector_label}
\alias{build_tma_sector_label}
\title{Build a Human-Readable Terminal-Airspace Sector Label}
\usage{
build_tma_sector_label(bearing_from, bearing_to)
}
\arguments{
\item{bearing_from}{Start bearing in degrees.}

\item{bearing_to}{End bearing in degrees.}
}
\value{
A character vector.
}
\description{
Returns a human-readable sector identifier from the start and end bearings.
}
