% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_utils.R
\name{check_zip_content}
\alias{check_zip_content}
\title{Check the Contents of a ZIP Archive}
\usage{
check_zip_content(path, archive)
}
\arguments{
\item{path}{Path to the directory containing the ZIP file.}

\item{archive}{Name of the ZIP file.}
}
\value{
A data frame as returned by \code{\link[utils:unzip]{utils::unzip()}} with \code{list = TRUE}.
}
\description{
Check the Contents of a ZIP Archive
}
