% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode.R
\name{decode_apdf}
\alias{decode_apdf}
\title{Decode APDF Columns}
\usage{
decode_apdf(data, dictionary = apdf_dictionary())
}
\arguments{
\item{data}{A tibble containing source data.}

\item{dictionary}{A tibble with \code{SOURCE_NAME} and \code{TARGET_NAME}.}
}
\value{
A tibble with harmonised column names.
}
\description{
Renames source columns in an APDF extract to harmonised names according to a
dictionary.
}
