% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode.R
\name{decode_nm_flights}
\alias{decode_nm_flights}
\title{Decode NM Flight Table Columns}
\usage{
decode_nm_flights(data, dictionary = nm_flights_dictionary())
}
\arguments{
\item{data}{A tibble containing source data.}

\item{dictionary}{A tibble with \code{SOURCE_NAME} and \code{TARGET_NAME}.}
}
\value{
A tibble with harmonised column names.
}
\description{
Renames source columns in an NM flight table extract to harmonised names
according to a dictionary.
}
