% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_flights.R
\name{is_ectrl_member_state_airport}
\alias{is_ectrl_member_state_airport}
\title{Check Whether Airports Match the EUROCONTROL Member State Pattern}
\usage{
is_ectrl_member_state_airport(icao_vec, icao_pattern = NULL)
}
\arguments{
\item{icao_vec}{Character vector of ICAO airport indicators.}

\item{icao_pattern}{Regular expression defining the airport scope.}
}
\value{
A logical vector.
}
\description{
Check Whether Airports Match the EUROCONTROL Member State Pattern
}
