% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{is_eurocontrol_airport}
\alias{is_eurocontrol_airport}
\title{Check Whether Airports Match the EUROCONTROL Member State Lookup}
\usage{
is_eurocontrol_airport(
  icao_vec,
  prefixes = eurocontrol_member_state_prefixes()
)
}
\arguments{
\item{icao_vec}{Character vector of ICAO airport indicators.}

\item{prefixes}{Optional character vector of two-letter ICAO prefixes.}
}
\value{
A logical vector.
}
\description{
Check Whether Airports Match the EUROCONTROL Member State Lookup
}
