% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apdf.R
\name{list_apdf_archives}
\alias{list_apdf_archives}
\title{List Available APDF Archives}
\usage{
list_apdf_archives(path, glob = "*.zip")
}
\arguments{
\item{path}{Directory containing APDF archives.}

\item{glob}{Optional glob pattern used to filter the listed files.}
}
\value{
A character vector of file paths.
}
\description{
Lists files in a directory containing APDF archives.
}
