% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_flights.R
\name{list_nm_flight_archives}
\alias{list_nm_flight_archives}
\title{List Available NM Flight Archives}
\usage{
list_nm_flight_archives(path, glob = "*.zip")
}
\arguments{
\item{path}{Directory containing NM flight table archives.}

\item{glob}{Optional glob pattern used to filter the listed files.}
}
\value{
A character vector of file paths.
}
\description{
Lists files in a directory containing NM flight table extracts.
}
