% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apdf_throughput.R
\name{prepare_ordered_throughput}
\alias{prepare_ordered_throughput}
\title{Prepare Ordered Throughput Curves}
\usage{
prepare_ordered_throughput(throughput_loads, years = NULL)
}
\arguments{
\item{throughput_loads}{Output from
\code{\link[=prepare_throughput_load_characteristics]{prepare_throughput_load_characteristics()}}.}

\item{years}{Optional years to keep.}
}
\value{
A tibble with rank-ordered hourly throughput by airport-year.
}
\description{
Orders hourly throughput observations from highest to lowest per airport-year
to support ranked-throughput illustrations.
}
