% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apdf.R
\name{read_pbwg_airport_traffic_files}
\alias{read_pbwg_airport_traffic_files}
\title{Read PBWG Airport Traffic Files}
\usage{
read_pbwg_airport_traffic_files(paths)
}
\arguments{
\item{paths}{Character vector of file paths.}
}
\value{
A tibble.
}
\description{
Reads one or more annual PBWG airport traffic files and returns a single
combined tibble.
}
