% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_flights.R
\name{read_pbwg_network_traffic_files}
\alias{read_pbwg_network_traffic_files}
\title{Read PBWG Network Traffic Files}
\usage{
read_pbwg_network_traffic_files(paths)
}
\arguments{
\item{paths}{Character vector of file paths.}
}
\value{
A tibble.
}
\description{
Reads one or more annual PBWG network traffic files and returns a single
combined tibble.
}
