% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apdf_punctuality.R
\name{read_pbwg_punctuality_files}
\alias{read_pbwg_punctuality_files}
\title{Read PBWG Punctuality Files}
\usage{
read_pbwg_punctuality_files(paths)
}
\arguments{
\item{paths}{Character vector of file paths.}
}
\value{
A tibble.
}
\description{
Reads one or more PBWG punctuality files and returns a single combined tibble.
}
