% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apdf_throughput.R
\name{read_pbwg_throughput_files}
\alias{read_pbwg_throughput_files}
\title{Read PBWG Throughput Files}
\usage{
read_pbwg_throughput_files(paths)
}
\arguments{
\item{paths}{Character vector of file paths.}
}
\value{
A tibble.
}
\description{
Reads one or more PBWG throughput files and returns a single combined tibble.
}
