% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apdf_tma.R
\name{read_pbwg_tma_files}
\alias{read_pbwg_tma_files}
\title{Read PBWG Daily Terminal-Airspace Files}
\usage{
read_pbwg_tma_files(paths)
}
\arguments{
\item{paths}{Character vector of file paths.}
}
\value{
A tibble.
}
\description{
Reads one or more PBWG daily terminal-airspace files and returns a single
combined tibble.
}
