% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apdf_txxt.R
\name{read_pbwg_txxt_files}
\alias{read_pbwg_txxt_files}
\title{Read PBWG Daily Taxi-Time Files}
\usage{
read_pbwg_txxt_files(paths)
}
\arguments{
\item{paths}{Character vector of file paths.}
}
\value{
A tibble.
}
\description{
Reads one or more PBWG daily taxi-time files and returns a single combined
tibble.
}
