% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apdf_throughput.R
\name{summarise_load_indices}
\alias{summarise_load_indices}
\title{Summarise Base-Load and Peak-Load Indices}
\usage{
summarise_load_indices(throughput_loads)
}
\arguments{
\item{throughput_loads}{Output from
\code{\link[=prepare_throughput_load_characteristics]{prepare_throughput_load_characteristics()}}.}
}
\value{
A tibble with total observed hours and BLI/PLI shares per
airport-year.
}
\description{
Calculates BLI and PLI shares per airport-year from hourly throughput load
characteristics.
}
