% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apdf_txxt.R
\name{txxt_reference_variants}
\alias{txxt_reference_variants}
\title{Taxi-Time Reference Variants}
\usage{
txxt_reference_variants()
}
\value{
A tibble describing the available variants.
}
\description{
Returns the supported taxi-time reference variants and their descriptions.
}
